+++ ckine/MBmodel.py
@@ -149,19 +149,44 @@ def runFullModel(x=False, time=[0.5], saveDict=False, singleCell=False):
         return masterSTAT
 
 
-# Tetra valent exploration functions below
+# CITEseq Tetra valent exploration functions below
 
-
-def cytBindingModel_bispec(counts, betaAffs, recXaff, val, mut, x=False):
+def cytBindingModel_CITEseq(counts, betaAffs, val, mut, x=False, date=False):
     """Runs binding model for a given mutein, valency, dose, and cell type."""
-  
+    
     doseVec = np.array([0.1])
-
-    recXaff = np.power(10, recXaff)
-
     recCount = np.ravel(counts)
 
-    mutAffDF = pd.read_csv(join(path_here, "data/WTmutAffData.csv"))
+    mutAffDF = pd.read_csv(join(path_here, "ckine/data/WTmutAffData.csv"))
+    Affs = mutAffDF.loc[(mutAffDF.Mutein == mut)]
+
+    Affs = np.power(np.array([Affs["IL2RaKD"].values, [betaAffs]]) / 1e9, -1)
+
+    Affs = np.reshape(Affs, (1, -1))
+    Affs = np.repeat(Affs, 2, axis=0)
+    np.fill_diagonal(Affs, 1e2)  # Each cytokine can only bind one a and one b
+
+    if doseVec.size == 1:
+        doseVec = np.array([doseVec])
+    output = np.zeros(doseVec.size)
+
+    for i, dose in enumerate(doseVec):
+        if x:
+            output[i] = polyc(dose / 1e9, np.power(10, x[0]), recCount, [[val, val]], [1.0], Affs)[0][1]
+        else:
+            output[i] = polyc(dose / 1e9, getKxStar(), recCount, [[val, val]], [1.0], Affs)[0][1]
+
+    return output
+
+
+def cytBindingModel_bispecCITEseq(counts, betaAffs, recXaff, val, mut, x=False):
+    """Runs bispecific binding model built for CITEseq data for a given mutein, epitope, valency, dose, and cell type."""
+  
+    recXaff = np.power(10, recXaff)
+    doseVec = np.array([0.1])
+    recCount = np.ravel(counts)
+
+    mutAffDF = pd.read_csv(join(path_here, "ckine/data/WTmutAffData.csv"))
     Affs = mutAffDF.loc[(mutAffDF.Mutein == mut)]
     Affs = np.power(np.array([Affs["IL2RaKD"].values, [betaAffs]]) / 1e9, -1)
     Affs = np.reshape(Affs, (1, -1))
